+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 LATEST CHANGES
 ==============
 
+2023-01-24
+----------
+- The parameter `general.sdk_verbose` has been moved to `debug.sdk_verbose` and set to `0` as default.
+- Add new parameter `general.optional_opencv_calibration_file` to use custom OpenCV camera calibrations.
+
 2023-12-21
 ----------
 - Add [new tutorial](https://github.com/stereolabs/zed-ros2-examples/tree/master/tutorials/zed_robot_integration) to illustrate how to integrate one or more ZED cameras on a robot